Extract advice walker effect types
